Closes #600: Improve tree reorder algorithm for nested modules
The idea here was to reduce the number of times the script needs to
interact with the database. With this new code, there should be fewer
`save()`, `saveAsRoot()`, and `self::find()` calls happening. The
trade-off is that there is now an `all()` call, which could cause
issues if there is truly a massive amount of items (thousands). But
for most use-cases, this new algorithm should behave better than the
old one when there is a middling number of items (hundreds).